# docs/oas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QDash API",
    "summary": "QDash API",
    "description": "API for QDash",
    "contact": {
      "name": "QDash",
      "email": "oqtopus-team@googlegroups.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": [
          "task"
        ],
        "summary": "List all tasks",
        "description": "List all tasks.\n\nArgs:\n----\n    ctx (ProjectContext): The project context with user and project information.\n    backend (Optional[str]): Optional backend name to filter tasks by.\n\nReturns:\n-------\n    list[TaskResponse]: The list of tasks.",
        "operationId": "listTasks",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backend",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional backend name to filter tasks by",
              "title": "Backend"
            },
            "description": "Optional backend name to filter tasks by"
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/task-files/settings": {
      "get": {
        "tags": [
          "task-file"
        ],
        "summary": "Get task file settings",
        "description": "Get task file settings from config/settings.yaml.\n\nReturns\n-------\n    Task file settings including default backend",
        "operationId": "getTaskFileSettings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskFileSettings"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/task-files/backends": {
      "get": {
        "tags": [
          "task-file"
        ],
        "summary": "List available task file backends",
        "description": "List all available backend directories in calibtasks.\n\nReturns\n-------\n    List of backend names and paths",
        "operationId": "listTaskFileBackends",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTaskFileBackendsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/task-files/tree": {
      "get": {
        "tags": [
          "task-file"
        ],
        "summary": "Get file tree for a specific backend",
        "description": "Get file tree structure for a specific backend directory.\n\nArgs:\n----\n    backend: Backend name (e.g., \"qubex\", \"fake\")\n\nReturns:\n-------\n    File tree structure for the backend",
        "operationId": "getTaskFileTree",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backend",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backend"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskFileTreeNode"
                  },
                  "title": "Response Gettaskfiletree"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/task-files/content": {
      "get": {
        "tags": [
          "task-file"
        ],
        "summary": "Get task file content for viewing/editing",
        "description": "Get task file content for viewing/editing.\n\nArgs:\n----\n    path: Relative path from CALIBTASKS_BASE_PATH (e.g., \"qubex/one_qubit_coarse/check_rabi.py\")\n\nReturns:\n-------\n    File content and metadata",
        "operationId": "getTaskFileContent",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Gettaskfilecontent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "task-file"
        ],
        "summary": "Save task file content",
        "description": "Save task file content.\n\nArgs:\n----\n    request: Save file request with path and content\n\nReturns:\n-------\n    Success message",
        "operationId": "saveTaskFileContent",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTaskFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Savetaskfilecontent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/task-files/tasks": {
      "get": {
        "tags": [
          "task-file"
        ],
        "summary": "List all tasks in a backend",
        "description": "List all task definitions found in a backend directory.\n\nParses Python files to extract task names, types, and descriptions.\nResults are cached and invalidated when files are modified.\n\nArgs:\n----\n    backend: Backend name (e.g., \"qubex\", \"fake\")\n    sort_order: Sort order for tasks (\"type_then_name\", \"name_only\", \"file_path\")\n\nReturns:\n-------\n    List of task information",
        "operationId": "listTaskInfo",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backend",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backend"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTaskInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileNodeType": {
        "type": "string",
        "enum": [
          "file",
          "directory"
        ],
        "title": "FileNodeType",
        "description": "File node type enum."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListTaskFileBackendsResponse": {
        "properties": {
          "backends": {
            "items": {
              "$ref": "#/components/schemas/TaskFileBackend"
            },
            "type": "array",
            "title": "Backends"
          }
        },
        "type": "object",
        "required": [
          "backends"
        ],
        "title": "ListTaskFileBackendsResponse",
        "description": "Response model for listing task file backends."
      },
      "ListTaskInfoResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskInfo"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "ListTaskInfoResponse",
        "description": "Response model for listing task info."
      },
      "ListTaskResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "ListTaskResponse",
        "description": "Response model for a list of tasks."
      },
...
      "SaveTaskFileRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "path",
          "content"
        ],
        "title": "SaveTaskFileRequest",
        "description": "Request model for saving task file content."
      },
...
      "TaskFileBackend": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "title": "TaskFileBackend",
        "description": "Task file backend model."
      },
      "TaskFileSettings": {
        "properties": {
          "default_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Backend"
          },
          "default_view_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default View Mode"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "type": "object",
        "title": "TaskFileSettings",
        "description": "Task file settings model."
      },
      "TaskFileTreeNode": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "type": {
            "$ref": "#/components/schemas/FileNodeType"
          },
          "children": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TaskFileTreeNode"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Children"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "type"
        ],
        "title": "TaskFileTreeNode",
        "description": "Task file tree node model."
      },
...
      "TaskInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "class_name",
          "file_path"
        ],
        "title": "TaskInfo",
        "description": "Task information extracted from Python file."
      },
      "TaskResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "input_parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/InputParameterModel"
            },
            "type": "object",
            "title": "Input Parameters"
          },
          "output_parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/InputParameterModel"
            },
            "type": "object",
            "title": "Output Parameters"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "task_type",
          "input_parameters",
          "output_parameters"
        ],
        "title": "TaskResponse",
        "description": "Response model for a task."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}